00_Arachnida/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
piexif==1.1.3
pillow==10.3.0
requests==2.32.2
soupsieve==2.5
urllib3==2.2.1